Addressed the feedbacks
