+++ src/Controls/tests/DeviceTests/Elements/RadioButton/RadioButtonTests.Windows.cs
@@ -1,4 +1,7 @@
 ﻿using Microsoft.Maui.Handlers;
+using System.ComponentModel;
+using Xunit;
+
 
 namespace Microsoft.Maui.DeviceTests
 {
