+++ src/Controls/tests/DeviceTests/Elements/RadioButton/RadioButtonTests.cs
@@ -4,7 +4,6 @@ using Microsoft.Maui.Controls.Xaml;
 using Microsoft.Maui.Handlers;
 using Microsoft.Maui.Hosting;
 using Xunit;
-using System.ComponentModel;
 
 namespace Microsoft.Maui.DeviceTests
 {
